tests: split dumpers tests and add edtf query
